requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
pydantic
pandas
openpyxl
psycopg2-binary
xlsxwriter
python-dotenv
tensorflow
numpy
Scikit-learn
requests
pathlib
flask-cors